lesson-6-hw  - сделал дз к 6 уроку
