+++ rest-api.http
@@ -46,6 +46,18 @@ Cookie: XDEBUG_SESSION=start
 
 ###
 
+POST http://localhost:80/users/create
+Accept: application/json
+Cookie: XDEBUG_SESSION=start
+
+{
+"username": "admin",
+"first_name": "Админ",
+"last_name": "Рут"
+}
+
+###
+
 POST http://localhost:80/posts/create
 Accept: application/json
 Cookie: XDEBUG_SESSION=start
@@ -75,8 +87,8 @@ Accept: application/json
 Cookie: XDEBUG_SESSION=start
 
 {
-"user_uuid": "10373537-0805-4d7a-830e-22b481b4859c",
-"post_uuid": "ba7ce47b-80e7-4dfc-857c-fb097565a333"
+"user_uuid": "6159f29f-9f6d-4b01-a022-cb0519a11ddd",
+"post_uuid": "76299b9b-e5d3-4915-80d9-8a0e834b4a3c"
 }
 
 ###
