--- src/Actions/ActionInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace Akseonov\Php2\Actions;
-
-use Akseonov\Php2\http\Request;
-use Akseonov\Php2\http\Response;
-
-interface ActionInterface
-{
-    public function handle(Request $request): Response;
-
-
-}
