--- src/Actions/Likes/CreatePostLike.php
@@ -1,70 +1,0 @@
-<?php
-
-namespace Akseonov\Php2\Actions\Likes;
-
-use Akseonov\Php2\Actions\ActionInterface;
-use Akseonov\Php2\Blog\PostLike;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\PostLikesRepositoryInterface;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\PostsRepositoryInterface;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\UsersRepositoryInterface;
-use Akseonov\Php2\Blog\UUID;
-use Akseonov\Php2\Exceptions\HttpException;
-use Akseonov\Php2\Exceptions\InvalidArgumentException;
-use Akseonov\Php2\Exceptions\PostNotFoundException;
-use Akseonov\Php2\Exceptions\UserNotFoundException;
-use Akseonov\Php2\http\ErrorResponse;
-use Akseonov\Php2\http\Request;
-use Akseonov\Php2\http\Response;
-use Akseonov\Php2\http\SuccessfulResponse;
-
-class CreatePostLike implements ActionInterface
-{
-    public function __construct(
-        private readonly PostLikesRepositoryInterface $postLikesRepository,
-        private readonly PostsRepositoryInterface $postsRepository,
-        private readonly UsersRepositoryInterface $usersRepository
-    )
-    {
-    }
-
-    public function handle(Request $request): Response
-    {
-        try {
-            $userUuid = new UUID($request->jsonBodyField('user_uuid'));
-        } catch (HttpException | InvalidArgumentException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        try {
-            $postUuid = new UUID($request->jsonBodyField('post_uuid'));
-        } catch (HttpException | InvalidArgumentException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        try {
-            $user = $this->usersRepository->get($userUuid);
-        } catch (UserNotFoundException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        try {
-            $post = $this->postsRepository->get($postUuid);
-        } catch (PostNotFoundException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        $newPostLikeUuid = UUID::random();
-
-        $postLike = new PostLike(
-            $newPostLikeUuid,
-            $post,
-            $user
-        );
-
-        $this->postLikesRepository->save($postLike);
-
-        return new SuccessfulResponse([
-            'uuid' => (string)$newPostLikeUuid,
-        ]);
-    }
-}
