--- src/Actions/Posts/CreatePost.php
@@ -1,63 +1,0 @@
-<?php
-
-namespace Akseonov\Php2\Actions\Posts;
-
-use Akseonov\Php2\Actions\ActionInterface;
-use Akseonov\Php2\Blog\Post;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\PostsRepositoryInterface;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\UsersRepositoryInterface;
-use Akseonov\Php2\Blog\UUID;
-use Akseonov\Php2\Exceptions\HttpException;
-use Akseonov\Php2\Exceptions\InvalidArgumentException;
-use Akseonov\Php2\Exceptions\UserNotFoundException;
-use Akseonov\Php2\http\Auth\IdentificationInterface;
-use Akseonov\Php2\http\ErrorResponse;
-use Akseonov\Php2\http\Request;
-use Akseonov\Php2\http\Response;
-use Akseonov\Php2\http\SuccessfulResponse;
-
-class CreatePost implements ActionInterface
-{
-    public function __construct(
-        private readonly PostsRepositoryInterface $postsRepository,
-        private readonly UsersRepositoryInterface $usersRepository,
-//        private IdentificationInterface $identification,
-    )
-    {
-    }
-
-    public function handle(Request $request): Response
-    {
-//        $user = $this->identification->user($request);
-        try {
-            $authorUuid = new UUID($request->jsonBodyField('author_uuid'));
-        } catch (HttpException | InvalidArgumentException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        try {
-            $user = $this->usersRepository->get($authorUuid);
-        } catch (UserNotFoundException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        $newPostUuid = UUID::random();
-
-        try {
-            $post = new Post(
-                $newPostUuid,
-                $user,
-                $request->jsonBodyField('title'),
-                $request->jsonBodyField('text'),
-            );
-        } catch (HttpException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        $this->postsRepository->save($post);
-
-        return new SuccessfulResponse([
-            'uuid' => (string)$newPostUuid,
-        ]);
-    }
-}
