--- src/Actions/Users/CreateUser.php
@@ -1,47 +1,0 @@
-<?php
-
-namespace Akseonov\Php2\Actions\Users;
-
-use Akseonov\Php2\Actions\ActionInterface;
-use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\UsersRepositoryInterface;
-use Akseonov\Php2\Blog\User;
-use Akseonov\Php2\Blog\UUID;
-use Akseonov\Php2\Exceptions\HttpException;
-use Akseonov\Php2\http\ErrorResponse;
-use Akseonov\Php2\http\Request;
-use Akseonov\Php2\http\Response;
-use Akseonov\Php2\http\SuccessfulResponse;
-use Akseonov\Php2\Person\Name;
-
-class CreateUser implements ActionInterface
-{
-    public function __construct(
-        private readonly UsersRepositoryInterface $usersRepository
-    )
-    {
-    }
-
-    public function handle(Request $request): Response
-    {
-        try {
-            $newUserUuid = UUID::random();
-
-            $user = new User(
-                $newUserUuid,
-                $request->jsonBodyField('username'),
-                new Name(
-                    $request->jsonBodyField('first_name'),
-                    $request->jsonBodyField('last_name')
-                )
-            );
-        } catch (HttpException $exception) {
-            return new ErrorResponse($exception->getMessage());
-        }
-
-        $this->usersRepository->save($user);
-
-        return new SuccessfulResponse([
-            'uuid' => (string)$newUserUuid
-        ]);
-    }
-}
