+++ src/http/Actions/Posts/CreatePost.php
@@ -1,0 +1,69 @@
+<?php
+
+namespace Akseonov\Php2\http\Actions\Posts;
+
+use Akseonov\Php2\Blog\Post;
+use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\PostsRepositoryInterface;
+use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\UsersRepositoryInterface;
+use Akseonov\Php2\Blog\UUID;
+use Akseonov\Php2\Exceptions\HttpException;
+use Akseonov\Php2\Exceptions\InvalidArgumentException;
+use Akseonov\Php2\Exceptions\UserNotFoundException;
+use Akseonov\Php2\http\Actions\ActionInterface;
+use Akseonov\Php2\http\ErrorResponse;
+use Akseonov\Php2\http\Request;
+use Akseonov\Php2\http\Response;
+use Akseonov\Php2\http\SuccessfulResponse;
+use Psr\Log\LoggerInterface;
+
+class CreatePost implements ActionInterface
+{
+    public function __construct(
+        private readonly PostsRepositoryInterface $postsRepository,
+        private readonly UsersRepositoryInterface $usersRepository,
+//        private IdentificationInterface $identification,
+        private readonly LoggerInterface $logger,
+    )
+    {
+    }
+
+    public function handle(Request $request): Response
+    {
+        $this->logger->info('CreatePost action start');
+//        $user = $this->identification->user($request);
+        try {
+            $authorUuid = $request->jsonBodyField('author_uuid');
+        } catch (HttpException $exception) {
+            $this->logger->warning($exception->getMessage());
+            return new ErrorResponse($exception->getMessage());
+        }
+
+        try {
+            $user = $this->usersRepository->get(new UUID($authorUuid));
+        } catch (UserNotFoundException | InvalidArgumentException $exception) {
+            $this->logger->warning($exception->getMessage());
+            return new ErrorResponse($exception->getMessage());
+        }
+
+        $newPostUuid = UUID::random();
+
+        try {
+            $post = new Post(
+                $newPostUuid,
+                $user,
+                $request->jsonBodyField('title'),
+                $request->jsonBodyField('text'),
+            );
+        } catch (HttpException $exception) {
+            $this->logger->warning($exception->getMessage());
+            return new ErrorResponse($exception->getMessage());
+        }
+
+        $this->postsRepository->save($post);
+        $this->logger->info("CreatePost action save post: $newPostUuid");
+
+        return new SuccessfulResponse([
+            'uuid' => (string)$newPostUuid,
+        ]);
+    }
+}
