+++ tests/http/Actions/Posts/CreatePostActionTest.php
@@ -1,0 +1,253 @@
+<?php
+
+namespace Akseonov\Php2\UnitTests\http\Actions\Posts;
+
+use Akseonov\Php2\Blog\Post;
+use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\PostsRepositoryInterface;
+use Akseonov\Php2\Blog\Repositories\RepositoryInterfaces\UsersRepositoryInterface;
+use Akseonov\Php2\Blog\User;
+use Akseonov\Php2\Blog\UUID;
+use Akseonov\Php2\Exceptions\PostNotFoundException;
+use Akseonov\Php2\Exceptions\UserNotFoundException;
+use Akseonov\Php2\http\Actions\Posts\CreatePost;
+use Akseonov\Php2\http\ErrorResponse;
+use Akseonov\Php2\http\Request;
+use Akseonov\Php2\http\SuccessfulResponse;
+use Akseonov\Php2\Person\Name;
+use Akseonov\Php2\UnitTests\DummyLogger;
+use JsonException;
+use PHPUnit\Framework\TestCase;
+
+class CreatePostActionTest extends TestCase
+{
+    private function postsRepository(): PostsRepositoryInterface
+    {
+        return new class() implements PostsRepositoryInterface
+        {
+            private bool $called = false;
+
+            public function __construct()
+            {
+            }
+
+            public function save(Post $post): void
+            {
+                $this->called = true;
+            }
+
+            public function get(UUID $uuid): Post
+            {
+                throw new PostNotFoundException('Not found');
+            }
+
+            public function getByTitle(string $title): Post
+            {
+                throw new PostNotFoundException('Not found');
+            }
+
+            public function getCalled(): bool
+            {
+                return $this->called;
+            }
+
+            public function delete(UUID $uuid): void
+            {
+            }
+        };
+    }
+
+    private function usersRepository(array $users): UsersRepositoryInterface
+    {
+        return new class($users) implements UsersRepositoryInterface
+        {
+            public function __construct(
+                private array $users
+            )
+            {
+            }
+
+            public function save(User $user): void
+            {
+            }
+
+            public function get(UUID $uuid): User
+            {
+                foreach ($this->users as $user) {
+                    if ($user instanceof User && (string)$uuid === $user->getUuid()) {
+                        return $user;
+                    }
+                }
+                throw new UserNotFoundException('Not found');
+            }
+
+            public function getByUsername(string $username): User
+            {
+                throw new UserNotFoundException('Not found');
+            }
+        };
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsErrorResponseIfCannotDecodeJsonBody(): void
+    {
+        $request = new Request([], [], "");
+
+        $postsRepository = $this->postsRepository([]);
+        $usersRepository = $this->usersRepository([]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(ErrorResponse::class, $response);
+        $this->expectOutputString('{"success":false,"reason":"Cannot decode json body"}');
+
+        $response->send();
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsErrorResponseIfNoAuthorUuidProvided(): void
+    {
+        $request = new Request([], [], '{"uuid":"2342342"}');
+
+        $postsRepository = $this->postsRepository([]);
+        $usersRepository = $this->usersRepository([]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(ErrorResponse::class, $response);
+        $this->expectOutputString('{"success":false,"reason":"No such Field: author_uuid"}');
+
+        $response->send();
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsErrorResponseIfNoTitleProvided(): void
+    {
+        $request = new Request([], [], '{"author_uuid":"10373537-0805-4d7a-830e-22b481b4859c"}');
+
+        $postsRepository = $this->postsRepository([]);
+        $usersRepository = $this->usersRepository([
+            new User(
+                new UUID('10373537-0805-4d7a-830e-22b481b4859c'),
+                'ivan',
+                new Name('Ivan', 'Nikitin')
+            ),
+        ]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(ErrorResponse::class, $response);
+        $this->expectOutputString('{"success":false,"reason":"No such Field: title"}');
+
+        $response->send();
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsErrorResponseIfNoTextProvided(): void
+    {
+        $request = new Request([], [], '{"author_uuid":"10373537-0805-4d7a-830e-22b481b4859c","title":"title"}');
+
+        $postsRepository = $this->postsRepository([]);
+        $usersRepository = $this->usersRepository([
+            new User(
+                new UUID('10373537-0805-4d7a-830e-22b481b4859c'),
+                'ivan',
+                new Name('Ivan', 'Nikitin')
+            ),
+        ]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(ErrorResponse::class, $response);
+        $this->expectOutputString('{"success":false,"reason":"No such Field: text"}');
+
+        $response->send();
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsErrorResponseIfNotFoundUser(): void
+    {
+        $request = new Request([], [], '{"author_uuid":"10373537-0805-4d7a-830e-22b481b4859c","title":"title","text":"text"}');
+
+        $postsRepository = $this->postsRepository([]);
+        $usersRepository = $this->usersRepository([]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(ErrorResponse::class, $response);
+        $this->expectOutputString('{"success":false,"reason":"Not found"}');
+
+        $response->send();
+    }
+
+    /**
+     * @runInSeparateProcess
+     * @preserveGlobalState disabled
+     * @throws JsonException
+     */
+    public function testItReturnsSuccessfulResponse(): void
+    {
+        $request = new Request([], [], '{"author_uuid":"10373537-0805-4d7a-830e-22b481b4859c","title":"title","text":"text"}');
+
+        $postsRepository = $this->postsRepository([]);
+
+        $usersRepository = $this->usersRepository([
+            new User(
+                new UUID('10373537-0805-4d7a-830e-22b481b4859c'),
+                'username',
+                new Name('name', 'surname')
+            ),
+        ]);
+
+        $action = new CreatePost($postsRepository, $usersRepository, new DummyLogger());
+
+        $response = $action->handle($request);
+
+        $this->assertInstanceOf(SuccessfulResponse::class, $response);
+        $this->expectOutputString('{"success":true,"data":{"uuid":"351739ab-fc33-49ae-a62d-b606b7038c87"}}');
+        $this->setOutputCallback(function ($data){
+            $dataDecode = json_decode(
+                $data,
+                associative: true,
+                flags: JSON_THROW_ON_ERROR
+            );
+            var_dump($dataDecode);
+            $dataDecode['data']['uuid'] = "351739ab-fc33-49ae-a62d-b606b7038c87";
+            return json_encode(
+                $dataDecode,
+                JSON_THROW_ON_ERROR
+            );
+        });
+
+        $response->send();
+    }
+}
